requirements.txt
numpy
tensorflow==2.5
scikit-image
opencv-python
pillow
pandas
scikit_learn
